% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_tibble.R
\name{as_tibble}
\alias{as_tibble}
\title{Coerce to tibble. Enable preserving row names when coercing matrix
and time-series-like objects with row names to tibbles}
\usage{
as_tibble(x, preserve_row_names = FALSE, ...)
}
\arguments{
\item{x}{A list, matrix, xts, zoo, timeSeries, etc object.}

\item{preserve_row_names}{When \code{TRUE}, creates a row.names column with
names of rows.}

\item{...}{Additional parameters passed to the appropriate
\code{\link[tibble]{as_tibble}}
function.}
}
\value{
Returns a \code{tibble} object.
}
\description{
Coerce to tibble. Enable preserving row names when coercing matrix
and time-series-like objects with row names to tibbles
}
\details{
\code{tidyquant::as_tibble} is a wrapper for \code{\link[tibble]{as_tibble}}
that includes a \code{preserve_row_names} argument. When \code{preserve_row_names = TRUE},
a new column, \code{row.names}, is created during object coercion.

It is possible to convert \code{xts}, \code{zoo}, \code{timeSeries}, \code{ts},
and \code{irts} objects.
}
\examples{
# Load libraries
library(tidyquant)

m <- matrix(rnorm(50), ncol = 5)
colnames(m) <- c("a", "b", "c", "d", "e")
rownames(m) <- letters[1:nrow(m)]
m_tbl <- as_tibble(m, preserve_row_names = TRUE)

}

