% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R, R/tq_transmute.R
\name{deprecated}
\alias{deprecated}
\alias{tq_transform}
\alias{tq_transform_xy}
\title{Deprecated functions}
\usage{
tq_transform(data, ohlc_fun = OHLCV, transform_fun, col_rename = NULL, ...)

tq_transform_xy(data, x, y = NULL, transform_fun, col_rename = NULL, ...)
}
\arguments{
\item{data}{A \code{tibble} (tidy data frame) of data typically from \code{\link[=tq_get]{tq_get()}}.}

\item{ohlc_fun}{The \code{quantmod} function that identifies columns to pass to
the mutatation function. OHLCV is \code{quantmod} terminology for
open, high, low, close, and volume. Options include c(Op, Hi, Lo, Cl, Vo, Ad,
HLC, OHLC, OHLCV).}

\item{transform_fun}{Deprecated. Use \code{mutate_fun}.}

\item{col_rename}{A string or character vector containing names that can be used
to quickly rename columns.}

\item{...}{Additional parameters passed to the appropriate mutatation
function.}

\item{x}{Column names of variables to be passed to the mutatation
function (instead of OHLC functions).}

\item{y}{Column names of variables to be passed to the mutatation
function (instead of OHLC functions).}
}
\description{
A record of functions that have been deprecated.
}
\details{
\itemize{
\item \code{tq_transform()} - use \code{\link[=tq_transmute]{tq_transmute()}}
\item \code{tq_transform_xy()} - use \code{\link[=tq_transmute_xy]{tq_transmute_xy()}}
}
}
