% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel-stat-functions.R
\name{excel_stat_functions}
\alias{excel_stat_functions}
\alias{SUM}
\alias{AVERAGE}
\alias{MEDIAN}
\alias{MIN}
\alias{MAX}
\alias{COUNT}
\alias{COUNT_UNIQUE}
\alias{STDEV}
\alias{VAR}
\alias{COR}
\alias{COV}
\alias{FIRST}
\alias{LAST}
\alias{NTH}
\alias{ABS}
\alias{SQRT}
\alias{LOG}
\alias{EXP}
\alias{RETURN}
\alias{PCT_CHANGE}
\alias{CHANGE}
\alias{LAG}
\alias{LEAD}
\alias{CUMULATIVE_SUM}
\alias{CUMULATIVE_PRODUCT}
\alias{CUMULATIVE_MAX}
\alias{CUMULATIVE_MIN}
\alias{CUMULATIVE_MEAN}
\alias{CUMULATIVE_MEDIAN}
\title{Excel Statistical Functions}
\usage{
SUM(x)

AVERAGE(x)

MEDIAN(x)

MIN(x)

MAX(x)

COUNT(x)

COUNT_UNIQUE(x)

STDEV(x)

VAR(x)

COR(x, y)

COV(x, y)

FIRST(x)

LAST(x)

NTH(x, n = 1)

ABS(x)

SQRT(x)

LOG(x)

EXP(x)

RETURN(x, n = 1, fill_na = NA)

PCT_CHANGE(x, n = 1, fill_na = NA)

CHANGE(x, n = 1, fill_na = NA)

LAG(x, n = 1, fill_na = NA)

LEAD(x, n = 1, fill_na = NA)

CUMULATIVE_SUM(x)

CUMULATIVE_PRODUCT(x)

CUMULATIVE_MAX(x)

CUMULATIVE_MIN(x)

CUMULATIVE_MEAN(x)

CUMULATIVE_MEDIAN(x)
}
\arguments{
\item{x}{A vector. Most functions are designed for numeric data.
Some functions like \code{\link[=COUNT]{COUNT()}} handle multiple data types.}

\item{y}{A vector. Used in functions requiring 2 inputs.}

\item{n}{Values to offset. Used in functions like \code{\link[=LAG]{LAG()}}, \code{\link[=LEAD]{LEAD()}}, and \code{\link[=PCT_CHANGE]{PCT_CHANGE()}}}

\item{fill_na}{Fill missing (\code{NA}) values with a different value. Used in offsetting functions.}
}
\value{
\itemize{
\item \strong{Summary functions} return a single value
\item \strong{Mutation functions} return a mutated version of the vector
}
}
\description{
Common statistical functions familiar to users of Excel.

These functions are designed to help users coming from an \strong{Excel background}.
Most functions replicate the behavior of Excel:
\itemize{
\item Names in most cases match Excel function names
\item By default, missing values are ignored (same as in Excel)
}
}
\section{Useful functions}{


\strong{Summary Functions} - Return a single value from a vector
\itemize{
\item Sum: \code{\link[=SUM]{SUM()}}
\item Center: \code{\link[=AVERAGE]{AVERAGE()}}, \code{\link[=MEDIAN]{MEDIAN()}}
\item Spread: \code{\link[=STDEV]{STDEV()}}, \code{\link[=VAR]{VAR()}}
\item Range: \code{\link[=MIN]{MIN()}}, \code{\link[=MAX]{MAX()}}
\item Count: \code{\link[=COUNT]{COUNT()}}, \code{\link[=COUNT_UNIQUE]{COUNT_UNIQUE()}}
\item Position: \code{\link[=FIRST]{FIRST()}}, \code{\link[=LAST]{LAST()}}, \code{\link[=NTH]{NTH()}}
\item Correlation: \code{\link[=COR]{COR()}}, \code{\link[=COV]{COV()}}
}

\strong{Mutation Functions} - Transforms a vector
\itemize{
\item Transformation: \code{\link[=ABS]{ABS()}}, \code{\link[=SQRT]{SQRT()}}, \code{\link[=LOG]{LOG()}}, \code{\link[=EXP]{EXP()}}
\item Lags & Change (Offsetting Functions): \code{\link[=CHANGE]{CHANGE()}}, \code{\link[=PCT_CHANGE]{PCT_CHANGE()}}, \code{\link[=LAG]{LAG()}}, \code{\link[=LEAD]{LEAD()}}
\item Cumulative Totals: \code{\link[=CUMULATIVE_SUM]{CUMULATIVE_SUM()}}, \code{\link[=CUMULATIVE_PRODUCT]{CUMULATIVE_PRODUCT()}}
}
}

\examples{
# Libraries
library(tidyquant)
library(tidyverse)

# TODO

}
