% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-scale_manual.R
\name{palette_tq}
\alias{palette_tq}
\alias{palette_light}
\alias{palette_dark}
\alias{palette_green}
\title{tidyquant palettes for use with scales}
\usage{
palette_light()

palette_dark()

palette_green()
}
\description{
These palettes are mainly called internally by tidyquant \verb{scale_*_tq()} functions.
}
\examples{
library(scales)
scales::show_col(palette_light())

}
