% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tq_stock_list.R
\name{parse_dollar}
\alias{parse_dollar}
\title{Parse a string such as 23B or 100M into a number}
\usage{
parse_dollar(x)
}
\arguments{
\item{x}{A string in the form \code{$(number)(B|M|K)}}
}
\description{
Parse a string such as 23B or 100M into a number
}
