% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyquant.R
\docType{package}
\name{tidyquant}
\alias{tidyquant}
\alias{tidyquant-package}
\title{tidyquant: Bridging the gap between quantitative analysis with quantmod
and the data management workflow of the tidyverse}
\description{
The main advantage of \code{tidyquant} is to
bridge the gap between the best quantitative resources for collecting and
manipulating quantitative data, \code{quantmod}, \code{TTR}, and
\code{PerformanceAnalytics}, and the data modeling workflow
infrastructure of the \code{tidyverse}.
}
\details{
In this package, \code{tidyquant} functions and supporting data sets are
provided to seamlessly combine tidy tools with existing quantitative
analytics packages. The main advantage is being able to use tidy
functions with purrr for mapping and tidyr for nesting to extend modeling to
many stocks. See the tidyquant website for more information, documentation
and examples.

To learn more about tidytext, start with the vignettes:
 \code{browseVignettes(package = "tidyquant")}
}

