% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tq_performance.R
\name{tq_performance}
\alias{tq_performance}
\alias{tq_performance_}
\alias{tq_performance_fun_options}
\title{Aggregates a group of returns by security into performance returns}
\usage{
tq_performance(data, Ra, Rb = NULL, performance_fun, ...)

tq_performance_(data, Ra, Rb = NULL, performance_fun, ...)

tq_performance_fun_options()
}
\arguments{
\item{data}{A \code{tibble} (tidy data frame) of returns in tidy format (i.e long format).}

\item{Ra}{The column of asset returns}

\item{Rb}{The column of baseline returns (for functions that require comarison to a baseline)}

\item{performance_fun}{The performance function from \code{PerformanceAnalytics}. See
\code{tq_performance_fun_options()} for a complete list of integrated functions.}

\item{...}{Additional parameters passed to the \code{PerformanceAnalytics} function.}
}
\value{
Returns data in the form of a \code{tibble} object.
}
\description{
Aggregates a group of returns by security into performance returns
}
\details{
\code{tq_performance} is a wrapper for various \code{PerformanceAnalytics} functions
that return portfolio statistics.
The main advantage is the ability to scale with the \code{tidyverse}.

\code{Ra} and \code{Rb} are the columns containing asset and baseline returns, respectively.
These columns are mapped to the \code{PerformanceAnalytics} functions.

\code{tq_performance_fun_options} returns a list of \code{PerformanceAnalytics} functions
that can be supplied to the \code{performance_fun} argument.
}
\examples{
# Load libraries
library(tidyquant)

# TBD

}
\seealso{
\itemize{
  \item \code{\link{tq_portfolio}} which can be used to aggregate period returns from
  multiple stocks to period returns for a portfolio.
  \item The \code{PerformanceAnalytics} package, which is contains the underlying functions
  for the \code{performance_fun} argument. Additional parameters can be passed via \code{...}.
}
}
