% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tq_transform.R
\name{tq_transform}
\alias{tq_transform}
\alias{tq_transform_}
\title{Transforms quantitative data with results in \code{tibble} format}
\usage{
tq_transform_(x, col, transform, ...)

tq_transform(x, col, transform, ...)
}
\arguments{
\item{x}{A \code{tibble} (tidy data frame) of data from \code{\link{tq_get}}.}

\item{col}{The column to use during transformation}

\item{transform}{A character string representing the type of transformation
to perform. Options include "lag", "next", "delt", "first"
"last", "to.period", "period.apply", and "period.return".}

\item{...}{Additional parameters passed to the appropriate \code{quantmod}
function.}
}
\value{
Returns data in the form of a \code{tibble} object.
}
\description{
Transforms quantitative data with results in \code{tibble} format
}
\details{
\code{tq_transform} is a wrapper for various \code{quantmod} functions
that returns the results as a \code{tibble}. The main advantage is the
function can be used with \code{purrr}, \code{tidyr}, and \code{dplyr} verbs.
}
\examples{
# Load libraries
library(tidyverse)
library(tidyquant)

##### Basic Functionality

fb_stock_prices  <- tq_get("FB", get = "stock.prices")

# lag


# next


# delt

# first


# last


# to.period


# period.apply


# period.return
fb_stock_prices \%>\%
    tq_transform(col = adjusted, transform = "period.return", k = 1)


##### Non-standard evaluation: Programming with tq_tranform_()
col_var <- "adjusted"
tq_transform_(fb_stock_prices, col = col_var, transform = "period.return")
}

