% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{tq_transform}
\alias{tq_transform_xy}
\title{Deprecated functions}
\usage{
tq_transform(data, ohlc_fun = OHLCV, mutate_fun, col_rename = NULL, ...)

tq_transform_xy(data, x, y = NULL, mutate_fun, col_rename = NULL, ...)
}
\arguments{
\item{data}{A \code{tibble} (tidy data frame) of data typically from \code{\link[=tq_get]{tq_get()}}.}

\item{ohlc_fun}{Deprecated. Use \code{select}.}

\item{mutate_fun}{The mutation function from either the \code{xts},
\code{quantmod}, or \code{TTR} package. Execute \code{tq_mutate_fun_options()}
to see the full list of options by package.}

\item{col_rename}{A string or character vector containing names that can be used
to quickly rename columns.}

\item{...}{Additional parameters passed to the appropriate mutatation
function.}

\item{x}{Parameters used with \verb{_xy} that consist of column names of variables
to be passed to the mutatation function (instead of OHLC functions).}

\item{y}{Parameters used with \verb{_xy} that consist of column names of variables
to be passed to the mutatation function (instead of OHLC functions).}
}
\description{
A record of functions that have been deprecated.
}
\details{
\itemize{
\item \code{tq_transform()} - use \code{\link[=tq_transmute]{tq_transmute()}}
\item \code{tq_transform_xy()} - use \code{\link[=tq_transmute_xy]{tq_transmute_xy()}}
\item \code{as_xts()} - use \code{\link[timetk:tk_xts]{timetk::tk_xts()}}
\item \code{as_tibble()} - use \code{\link[timetk:tk_tbl]{timetk::tk_tbl()}}
\item \code{summarise_by_time()} - Moved to \code{timetk} package. Use \code{\link[timetk:summarise_by_time]{timetk::summarise_by_time()}}
}
}
