% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-quandl.R
\name{quandl_api_key}
\alias{quandl_api_key}
\title{Query or set Quandl API Key}
\usage{
quandl_api_key(api_key)
}
\arguments{
\item{api_key}{Optionally passed parameter to set Quandl \code{api_key}.}
}
\value{
Returns invisibly the currently set \code{api_key}
}
\description{
Query or set Quandl API Key
}
\details{
A wrapper for \code{Quandl::Quandl.api_key()}
}
\examples{

\dontrun{
quandl_api_key(api_key = "foobar")
}
}
\seealso{
\code{\link[=tq_get]{tq_get()}} \code{get = "quandl"}
}
