% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-tiingo.R
\name{tiingo_api_key}
\alias{tiingo_api_key}
\title{Set Tiingo API Key}
\usage{
tiingo_api_key(api_key)
}
\arguments{
\item{api_key}{Optionally passed parameter to set Tiingo \code{api_key}.}
}
\value{
Returns invisibly the currently set \code{api_key}
}
\description{
Set Tiingo API Key
}
\details{
A wrapper for \code{riingo::ringo_set_token()}
}
\examples{

\dontrun{
tiingo_api_key(api_key = "foobar")
}

}
\seealso{
\code{\link[=tq_get]{tq_get()}} \code{get = "tiingo"}
}
