% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tq_get.R
\name{tq_get}
\alias{tq_get}
\title{Get quantitative data in \code{tibble} format}
\usage{
tq_get(x, get = "stock.prices", ...)
}
\arguments{
\item{x}{A character string representing a single stock index,
stock symbol, metal symbol, currency combination, FRED code, etc.}

\item{get}{A character string representing the type of data to get
for \code{x}. Options include:
\itemize{
  \item \code{"stock.index"}: Get a list of all stocks in an index or exchange
  from \href{http://www.marketvolume.com/indexes_exchanges/}{marketvolume}.
  Set \code{x = "options"} to get the full list.
  \item \code{"stock.prices"}: Get the stock prices for a stock symbol from
  \href{https://finance.yahoo.com/}{Yahoo}.
  \item \code{"divs.splits"}: Get the dividends and splits for a stock symbol
  from \href{https://finance.yahoo.com/}{Yahoo}.
  \item \code{"financials"}: Get the income, balance sheet, and cash flow
  financial statements for a stock symbol from
  \href{https://www.google.com/finance}{Google}.
  \item \code{"metal.prices"}: Get the metal prices from
  \href{https://www.oanda.com/}{Oanda}.
  \item \code{"exchange.rates"}: Get exchange rates from
  \href{https://www.oanda.com/currency/converter/}{Oanda}.
  \item \code{"economic.data"}: Get economic data from
  \href{https://fred.stlouisfed.org/}{FRED}.
}}

\item{...}{Additional parameters passed to the appropriate \code{quantmod} or
\code{TTR} function. Popular optional parameters include:
\itemize{
  \item \code{from}: Optional. A character string representing a start date in
  YYYY-MM-DD format. No effect on \code{get = "stock.index"} or
  \code{get = "financials"}.
  \item \code{to}: A character string representing a end date in
  YYYY-MM-DD format. No effect on \code{get = "stock.index"} or
  \code{get = "financials"}.
  \item \code{adjust}: Affects \code{get = stock.prices} only. A boolean
  representing whether stock prices should be
  adjusted for dividends and splits. Set to \code{TRUE} by default.
}}
}
\value{
Returns data in the form of a \code{tibble} object.
}
\description{
Get quantitative data in \code{tibble} format
}
\details{
\code{tq_get} is a consolidated function that gets data from various
web sources. The function is a wrapper for several \code{quantmod} and \code{TTR}
functions. The results are always returned as a \code{tibble}. The advantages
are (1) only one function is needed for all data sources and (2) the function
can be seemlessly used with the tidyverse: \code{purrr}, \code{tidyr}, and
\code{dplyr} verbs.
}
\examples{
# Load libraries
library(tidyverse)
library(tidyquant)

##### Basic Functionality
# Get a stock list
# index_list <- c(
#     "DOWJONES", "DJI", "DJT", "DJU","SP100", "SP400", "SP500", "SP600",
#     "RUSSELL1000", "RUSSELL2000", "RUSSELL3000", "AMEX", "AMEXGOLD",
#     "AMEXOIL", "NASDAQ", "NASDAQ100", "NYSE", "SOX"
#      )
tq_get("SP500", get = "stock.index")

# Get stock prices for a stock; TTR::getYahooData()
aapl_stock_prices <- tq_get("AAPL")
cvx_stock_prices  <- tq_get("CVX", get = "stock.prices", adjust = FALSE,
                            from = "2014-01-01", to = "2015-01-01")

# Get dividends and splits for a stock
tq_get("AAPL", get = "divs.splits", from = "2016-01-01")

# Get financial statment data for a stock; quantmod::getFinancials()
appl_financials <- tq_get("AAPL", get = "financials")

# Get exchange rate data for a currency combindation; quantmod::getFX()
eur_usd <- tq_get("EUR/USD", get = "exchange.rates")

# Get metals data for a commodity; quantmod::getMetals()
plat_price_usd <- tq_get("plat", get = "metal.prices")
gold_price_eur <- tq_get("gold", get = "metal.prices",
                                 base.currency = "EUR")

# Get FRED economic data for a commodity code; quantmod::getSymbols.FRED()
# WTI crude oil spot prices
tq_get("DCOILWTICO", get = "economic.data")


##### Tidyverse functionality

# Combine dplyr::mutate and purrr::map() to get a nested tibble
# of historical stock prices from multiple stocks
FANG <- tibble(symbol = c("FB", "AMZN", "NFLX", "GOOGL")) \%>\%
    mutate(stock.prices = map(symbol, tq_get)) \%>\%
    unnest()

}

