% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tq_index.R
\name{tq_index}
\alias{tq_index}
\alias{tq_index_options}
\title{Get all stocks in a stock index in \code{tibble} format}
\usage{
tq_index(x, use_fallback = FALSE)

tq_index_options()
}
\arguments{
\item{x}{A single character string, a character vector or tibble representing a
single stock index or multiple stock indexes from www.marketvolume.com.}

\item{use_fallback}{A boolean that can be used to return a fallback data set
last downloaded when the package was updated. Useful if the website is down.
Set to \code{FALSE} by default.}
}
\value{
Returns data in the form of a \code{tibble} object.
}
\description{
Get all stocks in a stock index in \code{tibble} format
}
\details{
\code{tq_index()} returns the stock symbol and company name of every stock
in an index. Eighteen stock indexes are available from
\href{http://www.marketvolume.com/indexes_exchanges/}{www.marketvolume.com}.

\code{tq_index_options()} returns a list of stock indexes you can
choose from.
}
\examples{
# Load libraries
library(tidyquant)

##### Basic Functionality

# Get the list of `get` options
tq_index_options()

# Get all stocks in a stock index from www.marketvolume.com
tq_index("SP500")
}
\seealso{
\code{\link{tq_get}} to get stock prices, financials, key stats, etc.
}

